refactor(archive): move attachments out of lib
