+++ polylogue/lib/models.py
@@ -1,6 +1,6 @@
 """Canonical public root for conversation/message domain models."""
 
-from polylogue.lib.attachment.models import Attachment
+from polylogue.archive.attachment.models import Attachment
 from polylogue.lib.conversation.models import Conversation, ConversationSummary
 from polylogue.lib.message.models import DialoguePair, Message
